+++ samples/uptime_evt.py
@@ -40,7 +40,8 @@ def boot_time():
 def up_time(time_boot=None):
     if time_boot is None:
         time_boot = boot_time()
-    utime = datetime.datetime.now() - time_boot # noqa: DTZ005
+    utime = datetime.datetime.now(time_boot.tzinfo) - time_boot
+    # ignore milliseconds
     return datetime.timedelta(seconds=int(utime.total_seconds()))
 
 if __name__ == "__main__":